Fixed how we check if this is a booking request confirmation

+++ core-minicomponents/j02990add_tourist_tax.class.php
@@ -26,7 +26,7 @@ class j02990add_tourist_tax
         }
 
         //if this is a secret key payment (for an approved booking enquiry) we don`t need to add the tourist tax again
-		$sk = (int)jomresGetParam($_REQUEST, 'sk', '');
+		$sk = jomresGetParam($_REQUEST, 'sk', '');
 
         if ($sk != '') {
             return;
